feat(users): refactor user validation logic by creating dedicated validation functions for user queries and addresses refactor(users): remove redundant authorization checks and replace them with validation middleware for cleaner code chore(users): remove unused Joi imports and related code to streamline user route handling fix(general): add validateIsUserAdmin function to enforce admin role checks in middleware
